extract selectors to a separate module
